requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
umap-learn==0.5.6
hdbscan==0.8.33
Pillow==10.3.0
numpy==1.24.4
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.3.2
